Converted the Argon2Type pseudo-enum to a more pythonic style
Specifically, since in python flat is better than nested, I pulled the
enum out of the containing class and into the main namespace, this
allows usage with `argon_type=argon2.argon2_i` instead of the existing
`argon_type=argon2.Argon2Type.Argon2_i` which is unnecesarilly verbose.

This change also attemps to be more compliant with python best
practices, by converting to true camel case and converting the first
character of the enum to lowercase.

It maintains compatibility.